refactor: destructure mui theme definitions into multiple files
