+++ site/src/theme/components/MuiCheckbox.ts
@@ -1,0 +1,9 @@
+import { Components } from "@mui/material";
+
+export const MuiCheckbox: Components["MuiCheckbox"] = {
+  defaultProps: {
+    disableRipple: true,
+    disableFocusRipple: true,
+    disableTouchRipple: true,
+  },
+};
