+++ site/src/theme/components/MuiContainer.ts
@@ -1,0 +1,17 @@
+import { Components } from "@mui/material";
+
+export const MuiContainer: Components["MuiContainer"] = {
+  styleOverrides: {
+    root: {
+      "@media (min-width: 1200px)": {
+        maxWidth: "1400px",
+      },
+      paddingLeft: "16px",
+      paddingRight: "16px",
+      "@media (min-width): 900px": {
+        paddingLeft: "32px",
+        paddingRight: "32px",
+      },
+    },
+  },
+};
