+++ site/src/theme/components/MuiFormControl.ts
@@ -1,0 +1,9 @@
+import { Components } from "@mui/material";
+
+export const MuiFormControl: Components["MuiFormControl"] = {
+  styleOverrides: {
+    root: {
+      display: "block",
+    },
+  },
+};
