+++ site/src/theme/components/MuiIcon.ts
@@ -1,0 +1,7 @@
+import { Components } from "@mui/material";
+
+export const MuiIcon: Components["MuiIcon"] = {
+  defaultProps: {
+    baseClassName: "fas",
+  },
+};
