+++ site/src/theme/components/MuiSkeleton.ts
@@ -1,0 +1,10 @@
+import { Components } from "@mui/material";
+import { customColors } from "../palette";
+
+export const MuiSkeleton: Components["MuiSkeleton"] = {
+  styleOverrides: {
+    root: {
+      backgroundColor: customColors.gray["20"],
+    },
+  },
+};
