+++ site/src/theme/components/MuiTextField.ts
@@ -1,0 +1,13 @@
+import { Components } from "@mui/material";
+
+export const MuiTextField: Components["MuiTextField"] = {
+  defaultProps: {
+    InputLabelProps: {
+      disableAnimation: true,
+      shrink: true,
+    },
+    InputProps: {
+      notched: false,
+    },
+  },
+};
