+++ site/src/theme/components/index.ts
@@ -1,0 +1,48 @@
+import { ThemeOptions } from "@mui/material";
+
+import { MuiCssBaseline } from "./MuiCssBaseline";
+import { MuiTypography } from "./MuiTypography";
+import { MuiIconButton } from "./MuiIconButton";
+import { MuiPaper } from "./MuiPaper";
+import { MuiButton } from "./MuiButton";
+import { MuiContainer } from "./MuiContainer";
+import { MuiIcon } from "./MuiIcon";
+import { MuiSkeleton } from "./MuiSkeleton";
+import {
+  MuiListItemButton,
+  MuiListItemIcon,
+  MuiListItemText,
+} from "./MuiListItem";
+import { MuiLink } from "./MuiLink";
+import { MuiCheckbox } from "./MuiCheckbox";
+import { MuiTabs, MuiTab } from "./MuiTabs";
+import { MuiBreadcrumbs } from "./MuiBreadcrumbs";
+import { MuiFormControl } from "./MuiFormControl";
+import { MuiInputLabel } from "./MuiInputLabel";
+import { MuiInputBase } from "./MuiInputBase";
+import { MuiOutlinedInput } from "./MuiOutlinedInput";
+import { MuiTextField } from "./MuiTextField";
+
+export const components: ThemeOptions["components"] = {
+  MuiCssBaseline,
+  MuiTypography,
+  MuiIconButton,
+  MuiPaper,
+  MuiButton,
+  MuiContainer,
+  MuiIcon,
+  MuiSkeleton,
+  MuiListItemButton,
+  MuiListItemIcon,
+  MuiListItemText,
+  MuiLink,
+  MuiCheckbox,
+  MuiTabs,
+  MuiTab,
+  MuiBreadcrumbs,
+  MuiFormControl,
+  MuiInputLabel,
+  MuiInputBase,
+  MuiOutlinedInput,
+  MuiTextField,
+};
