+++ site/src/theme/index.ts
@@ -1,0 +1,13 @@
+import { createTheme } from "@mui/material";
+
+import { palette } from "./palette";
+import { typography } from "./typography";
+import { shadows } from "./shadows";
+import { components } from "./components";
+
+export const theme = createTheme({
+  palette,
+  typography,
+  shadows,
+  components,
+});
