+++ site/src/theme/typography.ts
@@ -1,0 +1,90 @@
+import { ThemeOptions } from "@mui/material";
+import { customColors } from "./palette";
+import { defaultTheme } from "./util";
+
+const FALLBACK_FONTS = [`"Helvetica"`, `"Arial"`, "sans-serif"];
+
+export const typography: ThemeOptions["typography"] = {
+  fontFamily: [`"Inter"`, ...FALLBACK_FONTS].join(", "),
+  fontSize: 16,
+  bpTitle: {
+    fontFamily: [`"Apercu Pro"`, ...FALLBACK_FONTS].join(", "),
+    fontSize: "var(--step-6)",
+    lineHeight: 1,
+    fontWeight: 700,
+    color: customColors.gray["80"],
+  },
+  bpSubtitle: {
+    fontFamily: [`"Apercu Pro"`, ...FALLBACK_FONTS].join(", "),
+    fontSize: "var(--step-3)",
+    lineHeight: 1.1,
+    fontWeight: 200,
+    color: customColors.gray["80"],
+  },
+  bpHeading1: {
+    fontFamily: [`"Apercu Pro"`, ...FALLBACK_FONTS].join(", "),
+    fontSize: "var(--step-5)",
+    lineHeight: 1.1,
+    fontWeight: 700,
+    color: customColors.gray["80"],
+  },
+  bpHeading2: {
+    fontFamily: [`"Apercu Pro"`, ...FALLBACK_FONTS].join(", "),
+    fontSize: "var(--step-4)",
+    fontWeight: 400,
+    lineHeight: 1.2,
+    color: customColors.gray["80"],
+  },
+  bpHeading3: {
+    fontFamily: [`"Apercu Pro"`, ...FALLBACK_FONTS].join(", "),
+    fontSize: "var(--step-3)",
+    lineHeight: 1.1,
+    color: customColors.gray["70"],
+  },
+  bpHeading4: {
+    fontFamily: [`"Apercu Pro"`, ...FALLBACK_FONTS].join(", "),
+    fontSize: "var(--step-2)",
+    lineHeight: 1.1,
+    color: customColors.gray["70"],
+  },
+  bpSmallCaps: {
+    fontFamily: [`"Apercu Pro"`, ...FALLBACK_FONTS].join(", "),
+    fontSize: "var(--step--1)",
+    lineHeight: 1.3,
+    color: customColors.gray["70"],
+    textTransform: "uppercase",
+  },
+  bpLargeText: {
+    fontSize: "var(--step-1)",
+    lineHeight: 1.1,
+    color: customColors.gray["80"],
+  },
+  bpBodyCopy: {
+    fontSize: "var(--step-0)",
+    fontWeight: 400,
+    lineHeight: 1.7,
+    color: customColors.gray["80"],
+    maxWidth: "62ch",
+    "& a": {
+      fontWeight: 600,
+      transition: defaultTheme.transitions.create("color"),
+      color: customColors.purple[600],
+      "&:hover": {
+        color: customColors.purple["600"],
+      },
+    },
+    /** @todo: figure out how to type this */
+  } as any,
+  bpSmallCopy: {
+    fontWeight: 500,
+    fontSize: "var(--step--1)",
+    lineHeight: 1.5,
+    color: customColors.gray["80"],
+  },
+  bpMicroCopy: {
+    fontWeight: 500,
+    fontSize: "var(--step--2)",
+    lineHeight: 1.1,
+    color: customColors.gray["80"],
+  },
+};
