+++ site/src/theme/util.ts
@@ -1,0 +1,3 @@
+import { createTheme } from "@mui/material";
+
+export const defaultTheme = createTheme();
